Moved Fridge and Shop Routes into Seperate Files
