+++ back-end/app.js
@@ -10,97 +10,11 @@ const bodyParser = require("body-parser");
 app.use(bodyParser.json()); // decode JSON-formatted incoming POST data
 app.use(bodyParser.urlencoded({ extended: true })); // decode url-encoded incoming POST
 var request = require("request");
-const fridgeDataJSON = require("../front-end/src/data/fridgeMockData.json");
-const fridgeData = Object.entries(fridgeDataJSON[0]);
-const shopDataJSON = require("../front-end/src/data/shoppingListMockData.json");
-const shopData = Object.entries(shopDataJSON[0]);
+const MyFridgeRoutes = require("./MyFridge-Routes");
+const ShoppingListRoutes = require("./ShoppingList-Routes");
 
-// Get Fridge Data
-app.get("/fridgeData", (req, res) => {
-  res.json(fridgeData);
-});
-
-// Delete a Specific Fridge Item
-app.delete("/fridgeData/:id", (req, res) => {
-  const { id } = req.params;
-  let deleted = false;
-  for (let i = 0; i < fridgeData.length; i++) {
-    var removeIndex = fridgeData[i][1]
-      .map(function (item) {
-        return item.id.toString(); // Since id param is a string
-      })
-      .indexOf(id);
-    if (removeIndex !== -1) {
-      fridgeData[i][1].splice(removeIndex, 1);
-      deleted = true;
-    }
-  }
-  if (deleted) {
-    res.status(200).json(fridgeData);
-  } else {
-    res.status(200).json({ message: "Does not Exist" });
-  }
-});
-
-// Get Shopping List Data
-app.get("/shopData", (req, res) => {
-  res.status(200).json(shopData);
-});
-
-// Add Items to Fridge from Shopping List
-app.post("/addToFridge", (req, res) => {
-  let AddData = req.body;
-  for (let i = 0; i < AddData.length; i++) {
-    fridgeData[AddData[i].type][1].push(AddData[i]);
-  }
-  res.status(200).json(shopData);
-});
-
-// Delete Multiple Items from Shopping List After Adding to Fridge
-app.delete("/shopData", (req, res) => {
-  let AddData = req.body;
-  let deleted = false;
-  for (let i = 0; i < AddData.length; i++) {
-    const id = AddData[i].id; // Specific Item's Id
-    const type = AddData[i].type; // Specific Item's Food Type
-    var removeIndex = shopData[type][1]
-      .map(function (item) {
-        return item.id.toString(); // Since id param is a string
-      })
-      .indexOf(id);
-    if (removeIndex !== -1) {
-      shopData[type][1].splice(removeIndex, 1);
-      deleted = true;
-    }
-  }
-  if (deleted) {
-    res.status(200).json(shopData);
-  } else {
-    res.status(200).json({ message: "Does not Exist" });
-  }
-});
-
-// Delete a Specific Shopping Item
-app.delete("/shopData/:id", (req, res) => {
-  const { id } = req.params;
-  let deleted = false;
-  for (let i = 0; i < shopData.length; i++) {
-    var removeIndex = shopData[i][1]
-      .map(function (item) {
-        return item.id.toString(); // Since id param is a string
-      })
-      .indexOf(id);
-    if (removeIndex !== -1) {
-      shopData[i][1].splice(removeIndex, 1);
-      deleted = true;
-    }
-  }
-  if (deleted) {
-    res.status(200).json(shopData);
-  } else {
-    res.status(200).json({ message: "Does not Exist" });
-  }
-});
+app.use("/fridgeData", MyFridgeRoutes);
+app.use("/shopData", ShoppingListRoutes);
 
 app.get("/getRecipe", (req, res) => {
   //add    :name parameters later
