+++ front-end/src/ShoppingList/ShoppingList.js
@@ -42,7 +42,7 @@ const ShoppingListView = (props) => {
   // Adding Items to Fridge and Deleting from Shopping List
   const onAddToFridge = async () => {
     let AddData = compileAddToFridgeItems();
-    await axios.post("/addToFridge", AddData);
+    await axios.post("/shopData/addToFridge", AddData);
     await axios.delete("/shopData", { data: AddData }).then((res) => {
       setShopData(res.data);
     });
